P2P_chatting_room/requirement.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
packaging==24.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
